% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{fire data set.}
\format{
Data frame with 15 rows and 2 variables:
\describe{
 \item{DISTANCE}{X distance from fire station in miles}
 \item{DAMAGE}{Y Fire Damage 1000s of dollars}
}
}
\source{
\url{https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
fire
}
\description{
A data set collected by a fire insurance company, includes a sample of 15 recent fires in a suburb.
}
\keyword{datasets}
