% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybinomial.R
\name{mybin}
\alias{mybin}
\title{Simulates a binomial experiment}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of experiment iterations}

\item{n}{trial size}

\item{p}{probability of success}
}
\value{
True
}
\description{
mybin takes in a specified experiment iteration, number trials, and probability of success
}
\examples{
mybin(iter=1000,n=10, p=0.7) #1000 iterations

}
