% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot2}
\alias{myboot2}
\title{Produce bootstrap intervals}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{iterations}

\item{x}{sample distribution}

\item{fun}{function to be applied to sample}

\item{alpha}{to produce confidence interval}

\item{cx}{scaling plot text relative to default}

\item{...}{arguments}
}
\description{
myboot2 takes in a sample dist. applies a function and limits related to alpha.
}
\examples{
set.seed(39)
sam=rnorm(25,mean=25,sd=10)
myboot2(10000,x=sam,fun="mean",alpha=0.05,xlab="mean",col="Purple",cx=1.5)

}
