% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{Produce a normal curve with shaded region}
\usage{
mycltb(n = 5, iter = 10000, p = 0.5, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{iterations}

\item{p}{probability}

\item{...}{arguments}
}
\description{
myncurve takes in a mean,  std dev. and an a limit
}
\examples{
mycltb(n=5,iter=10000,p=0.5)

}
