% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Produce a normal curve with shaded region}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{region limit}
}
\value{
area
}
\description{
myncurve takes in a mean,  std dev. and an a limit
}
\examples{
myncurve(mu=10, sigma=5, a=6)

}
