% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{pareto}
\alias{pareto}
\title{Produce a Pareto diagram}
\usage{
pareto(x, mn = "Pareto barplot", ...)
}
\arguments{
\item{x}{A vector}

\item{mn}{main title}

\item{...}{arguments}
}
\description{
Pareto takes in a vector and produces a pareto plot.
}
\examples{
freq = c(15, 8, 63, 20)
RL=c("None", "Both", "LegsO", "WheelsO")
rob<-as.numeric(freq, RL)
l=rep(RL,freq)
pareto(l)

}
