% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tss.R
\name{TSS}
\alias{TSS}
\title{Obtain total sum of squares}
\usage{
TSS(df, column)
}
\arguments{
\item{df}{Data Frame}

\item{column}{column name}
}
\value{
Variable
}
\description{
TSS takes in two parameters, a dataframe and a column nd returns the Total Sum of Squares
}
\examples{
col1 <- c(5, 6, 7, 8)
col2 <- c(1, 8, 3, 4)
df.ex <- data.frame(col1, col2)
TSS(df.ex, col2)

}
