% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{produce plots to test central limit theory}
\usage{
mycltb(n = 5, iter = 10000, p = 0.5, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{iterations}

\item{p}{probability}

\item{...}{arguments}
}
\description{
mycltb takes in an iteration,  size and probability
}
\examples{
mycltb(n=5,iter=10000,p=0.5)

}
