% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STATSOU.package.R
\docType{package}
\name{STATSOU.package}
\alias{STATSOU.package}
\title{STATS OU Package}
\description{
Contains function for computing the total sum of squares
Contains function for producing a Pareto Diagram
Contains function for evaluating DDT data set.
Contains function for producing normal curve and area
}
\author{
Tristan Thrasher \email{tristan.thrasher-1@ou.edu}
}
